backend_api/requirements.txt
# https://www.djangoproject.com/
django
Pillow

# https://www.django-rest-framework.org/
djangorestframework
markdown
django-filter

# https://django-rest-framework-simplejwt.readthedocs.io/en/latest/getting_started.html
djangorestframework-simplejwt

asgiref==3.4.1
autopep8==1.6.0
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
pycodestyle==2.8.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5